requirements.txt
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
daff==1.4.2
dbt-extractor==0.6.0
deepdiff==7.0.1
Faker==37.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
leather==0.4.0
Logbook==1.5.3
MarkupSafe==3.0.2
mashumaro==3.16
more-itertools==10.7.0
msgpack==1.1.0
networkx==3.5
numpy==2.2.6
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
parsedatetime==2.6
pathspec==0.12.1
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
google-cloud-storage==2.10.0
urllib3==2.4.0
zipp==3.22.0
python-dotenv
